backend/requirements.txt
# Django and REST Framework
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-prometheus-metrics==0.2.1

# Database
psycopg2-binary==2.9.9
drizzle-orm==0.29.3  # Optional Drizzle ORM support

# Authentication
pyjwt==2.8.0
cryptography==41.0.7

# Celery
celery==5.3.6
redis==5.0.1

# Monitoring and Logging
sentry-sdk==1.39.1
prometheus-client==0.19.0
django-prometheus==2.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0

# Development
black==23.12.1
isort==5.13.2
flake8==6.1.0
pytest==7.4.4
pytest-django==4.7.0
